+++ config/test.ci.exs
@@ -3,4 +3,5 @@ use Mix.Config
 config :trans, Trans.TestRepo,
   username: "postgres",
   password: "postgres",
-  database: "trans_test"
+  database: "trans_test",
+  log: false