+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Update dependencies to avoid compilation warnings
 - Require Elixir 1.7 or newer
 - Migrate from CircleCI to GitHub Actions
+- Allow translating entire structs
 
 # 2.2.0 - 2020-02-01
 - Require Elixir 1.6 or newer