+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Migrate from CircleCI to GitHub Actions
 - Allow translating entire structs
 - Add translate!/3 function to raise if a translation does not exist
+- Allow saving translations into embedded_schemas
 
 # 2.2.0 - 2020-02-01
 - Require Elixir 1.6 or newer