+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Allow translating entire structs
 - Add translate!/3 function to raise if a translation does not exist
 - Allow saving translations into embedded_schemas
+- Improve docs
 
 # 2.2.0 - 2020-02-01
 - Require Elixir 1.6 or newer